new classes, new services, new repo
new classes, repo and services for all classes
new service for authentication
new cors config
new security config

+++ jo2024-backend/src/main/java/com/jtaraconat/jo2024backend/Exceptions/TicketNotFoundException.java
@@ -1,7 +1,7 @@
 package com.jtaraconat.jo2024backend.Exceptions;
 
 public class TicketNotFoundException extends RuntimeException{
-    public TicketNotFoundException(long id ) {
+    public TicketNotFoundException(int id ) {
         super("Could not found the ticket with id: " + id);
     }
 }
