+++ jo2024-backend/src/main/java/com/jtaraconat/jo2024backend/Repositories/TicketRepository.java
@@ -4,4 +4,4 @@ import com.jtaraconat.jo2024backend.Models.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface TicketRepository extends JpaRepository<Ticket, Long> {}
+public interface TicketRepository extends JpaRepository<Ticket, Integer> {}
